# shared/openapi.yaml
openapi: 3.0.3
info:
  title: ExplorArte API
  version: 1.0.0
  description: >
    REST contract for the ExplorArte platform (Sueños y Letras). Both the React
    web app and the Expo mobile app talk to this API through the shared ApiClient.
    Until a server exists they run against an in-memory mock that implements the
    same shapes defined here. Implement this contract in Python (FastAPI) or Java
    (Spring Boot); the @explorarte/shared seed module provides initial data.
servers:
  - url: http://localhost:8000
    description: Local backend
  - url: https://api.explorarte.example
    description: Production
...
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List users, optionally filtered by approval status (admin)
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserProfile' }
  /admin/users/{id}/approve:
    post:
      tags: [Admin]
      summary: Approve a pending teacher account (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
  /admin/users/{id}/reject:
    post:
      tags: [Admin]
      summary: Reject a pending teacher account (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
  /admin/users/{id}:
    delete:
      tags: [Admin]
      summary: Delete a teacher account and its Firebase identity (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '400': { description: Refused — the account is an administrator }
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, name, lastname, email, phone, institucion, ubicacion, role, status]
      properties:
        id: { type: string }
        name: { type: string }
        lastname: { type: string }
        email: { type: string }
        phone: { type: string }
        institucion: { type: string }
        ubicacion: { type: string }
        role: { type: string, enum: [teacher, admin] }
        status: { type: string, enum: [pending, approved, rejected] }
        photo: { type: string, nullable: true }
        profileCompleted:
          type: boolean
          description: >
            False while an invited account has not completed its profile. Drives
            the welcome notice; flips to true when the profile is saved with a
            name, a lastname and a location.
...